+++ src-plugin/Plugin/PluginConfig.cs
@@ -127,11 +127,14 @@ namespace K4Arenas
 		[JsonPropertyName("default-weapon-settings")]
 		public DefaultWeaponSettings DefaultWeaponSettings { get; set; } = new DefaultWeaponSettings();
 
+		[JsonPropertyName("allowed-weapon-prefs")]
+		public AllowedWeaponPreferences AllowedWeaponPreferences { get; set; } = new AllowedWeaponPreferences();
+
 		[JsonPropertyName("arena-math-overrides")]
 		public int ArenaMathOverrides { get; set; } = 0;
 
 		[JsonPropertyName("ConfigVersion")]
-		public override int Version { get; set; } = 5;
+		public override int Version { get; set; } = 6;
 	}
 
 	public sealed class CompatibilitySettings
@@ -234,6 +237,27 @@ namespace K4Arenas
 
 		[JsonPropertyName("default-pistol")]
 		public string? DefaultPistol { get; set; } = null;
+	}
+
+	public sealed class AllowedWeaponPreferences
+	{
+		[JsonPropertyName("rifle")]
+		public bool Rifle { get; set; } = true;
+
+		[JsonPropertyName("sniper")]
+		public bool Sniper { get; set; } = true;
+
+		[JsonPropertyName("smg")]
+		public bool SMG { get; set; } = true;
+
+		[JsonPropertyName("lmg")]
+		public bool LMG { get; set; } = true;
+
+		[JsonPropertyName("shotgun")]
+		public bool Shotgun { get; set; } = true;
+
+		[JsonPropertyName("pistol")]
+		public bool Pistol { get; set; } = true;
 	}
 
 	public sealed class DatabaseSettings