+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 -- 2024.12.23 - 1.5.5
 
 - feat: Added 3 new natives - IsAFK, FindOpponents & TerminateRoundIfPossible
+- update: When a player executes !afk, execute TerminateRoundIfPossible.
+- update: Added `allowed-weapon-prefs` to toggle showing weaponTypes in `!guns` menu.
 
 -- 2024.12.20 - 1.5.4